raspberrypi/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==2.2.3
openai==0.28.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyserial==3.5
python-dotenv==1.0.1
requests==2.32.3
RPi.GPIO==0.7.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
